// src/app/core/models/common/marketing-administrative-fee/marketing-administrative-fee-response.model.ts
export class MarketingAdministrativeFeeResponse {
  marketingAdministrativeFeeId: number = 0;

  feePercentage: number = 0;

  isCurrent: boolean = false;

  createdBy: number = 0;
  createdByName?: string;

  createdDate!: Date;

  updatedBy?: number;
  updatedByName?: string;

  updatedDate?: Date;
}